Rename local-config to keystatic.config
