+++ examples/next-app/app/keystatic/keystatic.tsx
@@ -1,6 +1,6 @@
 'use client';
 
 import { makePage } from '@keystatic/next/ui/app';
-import localConfig from '../../local-config';
+import localConfig from '../../keystatic.config';
 
 export default makePage(localConfig);
